comisiones en cronogramas, horas_semanales float, validacion live en prevalidacion
- Soporte comisiones en Cronogramas: visualizar [C1]/[C2] en calendarios,
  editar comision desde dialogs y data_editor con auto-guardado,
  color por comision en modo "Por materia", tabla resumen por comision
- Respetar comisiones pre-asignadas del cronograma al prevalidar
  (Rule 0 en _derive_comisiones)
- horas_semanales de int a float en MateriaDB, Materia y servicios
  para soportar valores como 17.5h (step 0.25 en selector)
- Auto-guardado de horas_semanales desde prevalidacion con toast
- Header de expander usa valor de DB (no widget) para evitar
  falso positivo al cambiar sin guardar
- Validacion de horas/comisiones movida debajo del data_editor
  para reflejar cambios en tiempo real
- Fix timezone UTC en calendario (offset 3h en seleccion de rangos)
- Fix coercion time string->time en data_editor de cronogramas
- Horas totales sin redondeo: 17.5h se muestra como 17.5h, no 18h

+++ src/database/models.py
@@ -167,7 +167,7 @@ class MateriaDB(SQLModel, table=True):
     nombre: str = Field(min_length=1)
     codigo_guarani: Optional[str] = Field(default=None)
     cupo: Optional[int] = Field(default=None, gt=0)
-    horas_semanales: Optional[int] = Field(default=None, gt=0)
+    horas_semanales: Optional[float] = Field(default=None, gt=0)
     periodo: str = Field(default="cuatrimestral")  # "anual" o "cuatrimestral"
     active: bool = Field(default=True)
     virtual: bool = Field(default=False)
